Fix error of UNMET OPTIONAL DEPENDENCY fsevents@1.2.4
nodemon depends on chokidar, which depends on fsevents. fsevents is for
Mac OS.  On Linux the dependencies are optional.

With nodemon installed as usual, when we run 'npm ls', a list of
dependency errors will show due to package-lock.json records the
dependency graph.

To solve the above dilemma, we install nodemon by temporarily preventing
npm from generating package-lock.json through the flag of
--no-package-lock.  But npm will not automatically prune node modules.